Mudança de versões e adicionei  os repositórios

+++ src/main/java/com/lavajato/LavaJato_2/controller/ControladorServicosLimpeza.java
@@ -20,7 +20,7 @@ public class ControladorServicosLimpeza {
         servicoLimpezaRepository.save(servico);
     }
     @GetMapping("/buscar/{id}")
-    public ServicoLimpeza buscarServico(@PathVariable Long id) {
+    public ServicoLimpeza buscarServico(@PathVariable Integer id) {
         return servicoLimpezaRepository.findById(id).orElse(null);
     }
